相關資料補登改成小數

+++ recycle-common/src/main/java/com/coderman/common/model/business/InputElement.java
@@ -4,6 +4,7 @@ import lombok.Data;
 
 import javax.persistence.Id;
 import javax.persistence.Table;
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -15,7 +16,7 @@ public class InputElement {
 
     private Long itemId;
 
-    private Integer value;
+    private BigDecimal value;
 
     private String validMonth;
